Change logic of grid creation
Add variable to css file to hold number of rows/columns for the createGrid function depending on user input (if any, if not uses the default number of squares to evaluation rows/columns).

This has corrected an error where a fully drawn out grid makes the grid disappear. I believe the issue was happening as the previous logic was to calculate number of squares on the fly and updated/replace the css styling for each square. Then the styling was getting overridden by the drawing functionality and essentially made the divs disappear.